Revision: linuxbios@linuxbios.org--devel/freebios--devel--2.0--patch-39 Creator:  Li-Ta Lo <ollie@lanl.gov>
TLA sucks again

This is the third time I try to commit only the emulator changes.
I hope this patch contains the emulator changes only.


git-svn-id: svn://svn.coreboot.org/coreboot/trunk@1955 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/devices/emulator/x86emu/prim_ops.c
@@ -1921,7 +1921,7 @@ Implements the IMUL instruction and side effects.
 void imul_long_direct(u32 *res_lo, u32* res_hi,u32 d, u32 s)
 {
 #ifdef  __HAS_LONG_LONG__
-    s64 res = (s32)d * (s32)s;
+    s64 res = (s64)d * (s64)s;
 
     *res_lo = (u32)res;
     *res_hi = (u32)(res >> 32);
